CPR-1216 update search response to use canonical objects
