+++ src/main/kotlin/uk/gov/justice/digital/hmpps/personrecord/api/model/search/PersonSearchResponse.kt
@@ -1,38 +1,48 @@
 package uk.gov.justice.digital.hmpps.personrecord.api.model.search
 
 import io.swagger.v3.oas.annotations.media.Schema
-import uk.gov.justice.digital.hmpps.personrecord.jpa.entity.AddressEntity
-import uk.gov.justice.digital.hmpps.personrecord.jpa.entity.AddressUsageEntity
-import uk.gov.justice.digital.hmpps.personrecord.jpa.entity.ContactEntity
-import uk.gov.justice.digital.hmpps.personrecord.jpa.entity.PseudonymEntity
+import uk.gov.justice.digital.hmpps.personrecord.api.model.canonical.CanonicalAddress
+import uk.gov.justice.digital.hmpps.personrecord.api.model.canonical.CanonicalAlias
+import uk.gov.justice.digital.hmpps.personrecord.api.model.search.SearchMatchStatus.Companion.toSearchStatus
+import uk.gov.justice.digital.hmpps.personrecord.jpa.entity.PersonEntity
 import uk.gov.justice.digital.hmpps.personrecord.jpa.entity.ReferenceEntity
-import uk.gov.justice.digital.hmpps.personrecord.model.types.AddressStatusCode
-import uk.gov.justice.digital.hmpps.personrecord.model.types.AddressUsageCode
-import uk.gov.justice.digital.hmpps.personrecord.model.types.ContactType
-import uk.gov.justice.digital.hmpps.personrecord.model.types.CountryCode
 import uk.gov.justice.digital.hmpps.personrecord.model.types.IdentifierType
-import uk.gov.justice.digital.hmpps.personrecord.model.types.SexCode
 import uk.gov.justice.digital.hmpps.personrecord.model.types.SourceSystemType
-import uk.gov.justice.digital.hmpps.personrecord.model.types.TitleCode
 import uk.gov.justice.digital.hmpps.personrecord.model.types.UUIDStatusType
 import uk.gov.justice.digital.hmpps.personrecord.model.types.UUIDStatusType.ACTIVE
 import java.time.LocalDate
-import java.time.ZonedDateTime
 
 data class PersonSearchResponse(val data: List<SearchData>)
 
 data class SearchData(
   val name: SearchName,
-  val aliases: List<SearchAlias>,
-  val addresses: List<SearchAddress>,
-  val identifiers: List<SearchReference>,
+  val aliases: List<CanonicalAlias>,
+  val addresses: List<CanonicalAddress>,
+  val identifiers: List<SearchIdentifier>,
   val sourceSystem: SourceSystemType,
   val status: SearchMatchStatus,
-  @field:Schema(
-    example = """[{"name":{"firstName":"John","middleNames":"John","lastName":"Doe"},"aliases":[],"addresses":[],"identifiers":[],"sourceSystem":"NOMIS","status":"HIGH_CONFIDENCE_MATCH"}]""",
-  )
+  @field:Schema(example = """[{"name":{"firstName":"asdfra","middleNames":"fuemsa","lastName":"bexeste","dateOfBirth":"1995-08-15"},"aliases":[{"firstName":"jerklqr","lastName":"xwkuett","middleNames":"qltzzyv","title":{"code":MR,"description":Mr},"sex":{"code":M,"description":Male}}],"addresses":[{"cprAddressId":"de75b1fc-9c93-4293-bc7e-6bf46a3175d5","noFixedAbode":false,"startDate":"1994-08-13","startDateTime":"1994-08-13T00:00:00","endDate":null,"endDateTime":null,"postcode":"CD9 5KL","subBuildingName":"subBuildingName","buildingName":"buildingName","buildingNumber":"5","thoroughfareName":"thoroughfareName","dependentLocality":null,"postTown":null,"county":null,"country":null,"countryCode":null,"uprn":null,"status":{"code":null,"description":null},"comment":null,"typeVerified":null,"usages":[],"contacts":[]},{"cprAddressId":"2e423922-cc3f-463b-a018-acdbd91f2f3d","noFixedAbode":false,"startDate":"1965-12-08","startDateTime":"1965-12-08T00:00:00","endDate":null,"endDateTime":null,"postcode":"XV9 2XO","subBuildingName":null,"buildingName":null,"buildingNumber":null,"thoroughfareName":null,"dependentLocality":null,"postTown":null,"county":null,"country":null,"countryCode":null,"uprn":null,"status":{"code":null,"description":null},"comment":null,"typeVerified":null,"usages":[],"contacts":[]}],"identifiers":[{"type":"CRO","value":"628285/81W","comment":null},{"type":"PNC","value":"1954/1844444B","comment":null}],"sourceSystem":"NOMIS","status":"HIGH_CONFIDENCE_MATCH","linkedRecords":[{"name":{"firstName":"wofqyvj","middleNames":null,"lastName":"osqqttb","dateOfBirth":"2017-01-12"},"aliases":[{"firstName":"jmjcsuw","lastName":"gpybpvf","middleNames":"zagknav","title":{"code":null,"description":null},"sex":{"code":null,"description":null}}],"addresses":[{"cprAddressId":"c611b4d4-900b-4ed5-8cce-f361704c62fa","noFixedAbode":true,"startDate":"2015-01-22","startDateTime":"2015-01-22T00:00:00","endDate":null,"endDateTime":null,"postcode":"ID8 6MV","subBuildingName":null,"buildingName":null,"buildingNumber":null,"thoroughfareName":null,"dependentLocality":null,"postTown":null,"county":null,"country":null,"countryCode":null,"uprn":null,"status":{"code":null,"description":null},"comment":null,"typeVerified":null,"usages":[],"contacts":[]},{"cprAddressId":"195dbe32-a1fc-4a33-b73a-6313eba394fa","noFixedAbode":false,"startDate":"1950-01-09","startDateTime":"1950-01-09T00:00:00","endDate":null,"endDateTime":null,"postcode":"UD8 4AG","subBuildingName":null,"buildingName":null,"buildingNumber":null,"thoroughfareName":null,"dependentLocality":null,"postTown":null,"county":null,"country":null,"countryCode":null,"uprn":null,"status":{"code":null,"description":null},"comment":null,"typeVerified":null,"usages":[],"contacts":[]}],"identifiers":[{"type":"CRO","value":"948174/19X","comment":null},{"type":"PNC","value":"2019/2379161H","comment":null}],"sourceSystem":"NOMIS","status":"HIGH_CONFIDENCE_MATCH"}]}]""")
   var linkedRecords: List<SearchData> = emptyList(),
-)
+) {
+  companion object {
+    fun from(personEntity: PersonEntity): SearchData {
+      val mainPseudonym = personEntity.getPrimaryName()
+      return SearchData(
+        name = SearchName(
+          firstName = mainPseudonym.firstName,
+          middleNames = mainPseudonym.middleNames,
+          lastName = mainPseudonym.lastName,
+          dateOfBirth = mainPseudonym.dateOfBirth,
+        ),
+        aliases = CanonicalAlias.from(personEntity) ?: emptyList(),
+        addresses = personEntity.addresses.map { CanonicalAddress.from(it) },
+        identifiers = personEntity.references.map { SearchIdentifier.from(it) },
+        sourceSystem = personEntity.sourceSystem,
+        status = personEntity.personKey!!.status.toSearchStatus(),
+      )
+    }
+  }
+}
 
 data class SearchName(
   val firstName: String?,
@@ -41,105 +51,13 @@ data class SearchName(
   val dateOfBirth: LocalDate?,
 )
 
-data class SearchAlias(
-  val firstName: String? = null,
-  val lastName: String? = null,
-  val middleNames: String? = null,
-  val titleCode: TitleCode? = null,
-  val dateOfBirth: LocalDate? = null,
-  val sexCode: SexCode? = null,
-) {
-  companion object {
-    fun from(pseudonymEntity: PseudonymEntity): SearchAlias = SearchAlias(
-      titleCode = pseudonymEntity.titleCode,
-      firstName = pseudonymEntity.firstName,
-      middleNames = pseudonymEntity.middleNames,
-      lastName = pseudonymEntity.lastName,
-      dateOfBirth = pseudonymEntity.dateOfBirth,
-      sexCode = pseudonymEntity.sexCode,
-    )
-  }
-}
-
-data class SearchAddress(
-  val noFixedAbode: Boolean? = null,
-  val startDate: ZonedDateTime? = null,
-  val endDate: ZonedDateTime? = null,
-  val postcode: String? = null,
-  val fullAddress: String? = null,
-  val subBuildingName: String? = null,
-  val buildingName: String? = null,
-  val buildingNumber: String? = null,
-  val thoroughfareName: String? = null,
-  val dependentLocality: String? = null,
-  val postTown: String? = null,
-  val county: String? = null,
-  val countryCode: CountryCode? = null,
-  val uprn: String? = null,
-  val comment: String? = null,
-  val contacts: List<SearchContact> = emptyList(),
-  var statusCode: AddressStatusCode? = null,
-  var usages: List<SearchAddressUsage> = emptyList(),
-  var typeVerified: Boolean? = null,
-) {
-  companion object {
-    fun from(addressEntity: AddressEntity) = SearchAddress(
-      postcode = addressEntity.postcode,
-      fullAddress = addressEntity.fullAddress,
-      startDate = addressEntity.startDate,
-      endDate = addressEntity.endDate,
-      noFixedAbode = addressEntity.noFixedAbode,
-      subBuildingName = addressEntity.subBuildingName,
-      buildingName = addressEntity.buildingName,
-      buildingNumber = addressEntity.buildingNumber,
-      thoroughfareName = addressEntity.thoroughfareName,
-      dependentLocality = addressEntity.dependentLocality,
-      postTown = addressEntity.postTown,
-      county = addressEntity.county,
-      countryCode = addressEntity.countryCode,
-      uprn = addressEntity.uprn,
-      comment = addressEntity.comment,
-      statusCode = addressEntity.statusCode,
-      typeVerified = addressEntity.isVerified,
-      usages = addressEntity.usages.map { SearchAddressUsage.from(it) },
-      contacts = addressEntity.contacts.map { SearchContact.from(it) },
-    )
-  }
-}
-
-data class SearchAddressUsage(
-  val addressUsageCode: AddressUsageCode,
-  val isActive: Boolean,
-) {
-  companion object {
-    fun from(addressUsageEntity: AddressUsageEntity) = SearchAddressUsage(
-      addressUsageCode = addressUsageEntity.usageCode,
-      isActive = addressUsageEntity.active,
-    )
-  }
-}
-
-data class SearchContact(
-  val type: ContactType,
-  val value: String? = null,
-  val extension: String? = null,
-) {
-  companion object {
-    fun from(contactEntity: ContactEntity) = SearchContact(
-      type = contactEntity.contactType,
-      value = contactEntity.contactValue,
-      extension = contactEntity.extension,
-    )
-  }
-}
-
-data class SearchReference(
+data class SearchIdentifier(
   val type: IdentifierType,
   val value: String? = null,
   val comment: String? = null,
 ) {
   companion object {
-    fun from(referenceEntity: ReferenceEntity) = SearchReference(
+    fun from(referenceEntity: ReferenceEntity) = SearchIdentifier(
       type = referenceEntity.identifierType,
       value = referenceEntity.identifierValue,
       comment = referenceEntity.comment,
