+++ src/test/kotlin/uk/gov/justice/digital/hmpps/personrecord/api/controller/search/PersonSearchControllerIntTest.kt
@@ -5,11 +5,11 @@ import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
 import org.springframework.http.HttpStatus
 import uk.gov.justice.digital.hmpps.personrecord.api.constants.Roles.API_SEARCH_ONLY
+import uk.gov.justice.digital.hmpps.personrecord.api.model.canonical.CanonicalAlias
 import uk.gov.justice.digital.hmpps.personrecord.api.model.search.PersonSearchRequest
 import uk.gov.justice.digital.hmpps.personrecord.api.model.search.PersonSearchResponse
-import uk.gov.justice.digital.hmpps.personrecord.api.model.search.SearchAlias
+import uk.gov.justice.digital.hmpps.personrecord.api.model.search.SearchIdentifier
 import uk.gov.justice.digital.hmpps.personrecord.api.model.search.SearchMatchStatus
-import uk.gov.justice.digital.hmpps.personrecord.api.model.search.SearchReference
 import uk.gov.justice.digital.hmpps.personrecord.client.model.match.PersonMatchScore
 import uk.gov.justice.digital.hmpps.personrecord.config.WebTestBase
 import uk.gov.justice.digital.hmpps.personrecord.model.types.NameType
@@ -75,8 +75,8 @@ class PersonSearchControllerIntTest : WebTestBase() {
       assertThat(strongestPersonFromResponse.name.dateOfBirth).isEqualTo(strongestPersonPrimaryPseudonym.dateOfBirth)
       assertThat(strongestPersonFromResponse.sourceSystem).isEqualTo(strongestMatchPersonEntity.sourceSystem)
       assertThat(strongestPersonFromResponse.status).isEqualTo(SearchMatchStatus.HIGH_CONFIDENCE_MATCH)
-      assertThat(strongestPersonFromResponse.aliases).usingRecursiveComparison().isEqualTo(strongestMatchPersonEntity.getAliases().map { SearchAlias.from(it) })
-      assertThat(strongestPersonFromResponse.identifiers).usingRecursiveComparison().isEqualTo(strongestMatchPersonEntity.references.map { SearchReference.from(it) })
+      assertThat(strongestPersonFromResponse.aliases).usingRecursiveComparison().isEqualTo(CanonicalAlias.from(strongestMatchPersonEntity))
+      assertThat(strongestPersonFromResponse.identifiers).usingRecursiveComparison().isEqualTo(strongestMatchPersonEntity.references.map { SearchIdentifier.from(it) })
       assertThat(strongestPersonFromResponse.addresses).hasSize(strongestMatchPersonEntity.addresses.size)
 
       val weakestPersonFromResponse = personSearchResponse.data.first().linkedRecords.first()
@@ -87,8 +87,8 @@ class PersonSearchControllerIntTest : WebTestBase() {
       assertThat(weakestPersonFromResponse.name.dateOfBirth).isEqualTo(weakestPersonPrimaryPseudonym.dateOfBirth)
       assertThat(weakestPersonFromResponse.sourceSystem).isEqualTo(weakestMatchPersonEntity.sourceSystem)
       assertThat(weakestPersonFromResponse.status).isEqualTo(SearchMatchStatus.HIGH_CONFIDENCE_MATCH)
-      assertThat(weakestPersonFromResponse.aliases).usingRecursiveComparison().isEqualTo(weakestMatchPersonEntity.getAliases().map { SearchAlias.from(it) })
-      assertThat(weakestPersonFromResponse.identifiers).usingRecursiveComparison().isEqualTo(weakestMatchPersonEntity.references.map { SearchReference.from(it) })
+      assertThat(weakestPersonFromResponse.aliases).usingRecursiveComparison().isEqualTo(CanonicalAlias.from(weakestMatchPersonEntity))
+      assertThat(weakestPersonFromResponse.identifiers).usingRecursiveComparison().isEqualTo(weakestMatchPersonEntity.references.map { SearchIdentifier.from(it) })
       assertThat(weakestPersonFromResponse.addresses).hasSize(weakestMatchPersonEntity.addresses.size)
     }
